requirements.txt
altair==5.3.0
annotated-types==0.6.0
anyio==3.7.1
asgiref==3.7.2
attrs==23.2.0
Babel==2.9.1
beautifulsoup4==4.12.2
blinker==1.7.0
cached-property==1.5.2
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
chardet==4.0.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==3.4.8
decorator==4.4.2
distro==1.9.0
Django==5.0.2
docopt==0.6.2
docutils==0.17
ebaysdk==2.1.5
exceptiongroup==1.2.1
freezegun==1.1.0
geographiclib==2.0
geoip2==2.9.0
geopy==2.4.1
gevent==21.8.0
gitdb==4.0.11
GitPython==3.1.43
greenlet==1.1.2
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==2.10
isodate==0.6.1
Jinja2==3.0.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
libsass==0.20.1
lxml==4.8.0
lyzr-automata==0.1.3
markdown-it-py==3.0.0
MarkupSafe==2.0.1
maxminddb==2.5.1
mdurl==0.1.2
num2words==0.5.10
numpy==1.26.4
ofxparse==0.21
openai==1.3.4
packaging==24.0
pandas==2.2.2
passlib==1.7.4
Pillow==9.0.0
platformdirs==4.1.0
polib==1.1.1
postgis==1.0.4
protobuf==4.25.3
psutil==5.9.0
psycopg2==2.9.2
psycopg2-binary==2.9.9
pyarrow==16.0.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.7.1
pydantic_core==2.18.2
pydeck==0.9.0b1
pydot==1.4.2
Pygments==2.17.2
pyOpenSSL==21.0.0
pyparsing==3.1.1
pypdf @ git+https://github.com/py-pdf/PyPDF2.git@4bdca1628606706f652f02c260c2e7dd53fb2ff7
PyPDF2==1.23
pyserial==3.5
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-ldap==3.4.0
python-stdnum==1.17
pytz==2023.3.post1
pyusb==1.2.1
qrcode==7.3.1
razorpay==1.4.1
referencing==0.35.0
reportlab==3.6.8
requests==2.31.0
requests-file==1.5.1
requests-toolbelt==1.0.0
rich==13.7.1
rjsmin==1.2.2
rpds-py==0.18.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.5
SpeechRecognition==3.10.3
sqlparse==0.4.4
streamlit==1.33.0
tenacity==8.2.3
toml==0.10.2
toolz==0.12.1
tornado==6.4
tqdm==4.66.2
typing_extensions==4.9.0
tzdata==2024.1
urllib3==1.26.5
vobject==0.9.6.1
Werkzeug==2.0.2
WooCommerce==3.0.0
xlrd==1.2.0
XlsxWriter==3.0.2
xlwt==1.3.0
zeep==4.1.0
zope.event==5.0
zope.interface==6.1
